+++ bitbucket/resource_repository_variable.go
@@ -2,6 +2,7 @@ package bitbucket
 
 import (
 	"bytes"
+	"crypto/sha256"
 	"encoding/json"
 	"fmt"
 	"io/ioutil"
@@ -10,6 +11,11 @@ import (
 
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
 )
+
+func hashValue(v string) string {
+	b := sha256.Sum256([]byte(v))
+	return fmt.Sprintf("%x", b)
+}
 
 // RepositoryVariable structure for handling key info
 type RepositoryVariable struct {
@@ -36,13 +42,24 @@ func resourceRepositoryVariable() *schema.Resource {
 				Required: true,
 			},
 			"value": {
-				Type:     schema.TypeString,
-				Required: true,
+				Type:      schema.TypeString,
+				Required:  true,
+				Sensitive: true,
+			},
+			"hash": {
+				Type:      schema.TypeString,
+				Computed:  true,
+				Sensitive: true,
 			},
 			"secured": {
 				Type:     schema.TypeBool,
 				Optional: true,
 				Default:  false,
+			},
+			"always_override": {
+				Type:     schema.TypeBool,
+				Optional: true,
+				Default:  true,
 			},
 			"repository": {
 				Type:     schema.TypeString,
@@ -90,6 +107,7 @@ func resourceRepositoryVariableCreate(d *schema.ResourceData, m interface{}) err
 		return decodeerr
 	}
 	d.Set("uuid", rv.UUID)
+	d.Set("hash", hashValue(rvcr.Value))
 	d.SetId(rv.Key)
 
 	return resourceRepositoryVariableRead(d, m)
@@ -119,8 +137,18 @@ func resourceRepositoryVariableRead(d *schema.ResourceData, m interface{}) error
 
 		d.Set("uuid", rv.UUID)
 		d.Set("key", rv.Key)
-		d.Set("value", rv.Value)
 		d.Set("secured", rv.Secured)
+		switch {
+		case rv.Secured && !d.Get("always_override").(bool):
+			v := d.Get("value").(string)
+			h := d.Get("hash").(string)
+			if hashValue(v) != h {
+				d.Set("value", "")
+				d.Set("hash", "")
+			}
+		default:
+			d.Set("value", rv.Value)
+		}
 	}
 
 	if rvReq.StatusCode == 404 {
@@ -152,6 +180,7 @@ func resourceRepositoryVariableUpdate(d *schema.ResourceData, m interface{}) err
 		return nil
 	}
 
+	d.Set("hash", hashValue(rvcr.Value))
 	return resourceRepositoryVariableRead(d, m)
 }